refactor: modular code
